+++ src/components/SiteHeader/SiteHeader.tsx
@@ -1,29 +1,100 @@
 import { FC } from "react";
 import Link from "next/link";
-import styled, { useTheme } from "styled-components";
+import { useTheme } from "styled-components";
+import { useRouter } from "next/router";
 
 import Flex from "../Flex";
-import P, { Heading } from "../Typography";
+import P from "../Typography";
 import FixedHeader from "../FixedHeader";
 import { Menu } from "../Menu";
 import Logo from "../Logo";
+import MenuLinks from "../MenuLinks";
 import { useMenuContext } from "../../context/Menu";
 import IconButton from "../Button/IconButton";
-import Icon from "../Icon";
-
-const Home = styled(Heading)`
-  opacity: 0.6;
-  text-decoration: underline;
-`;
-
-const LocationIcon = styled(Icon)`
-  opacity: 0.6;
-`;
+import { MenuListItemProps } from "../MenuLinks/types";
+import {
+  getHelpUrl,
+  getPupilsUrl,
+  getTeachersUrl,
+} from "../../common-lib/urls";
+import useAnalytics from "../../context/Analytics/useAnalytics";
 
 const SiteHeader: FC = () => {
   const theme = useTheme();
-
   const { toggleMenu } = useMenuContext();
+  const { asPath } = useRouter();
+  const { track } = useAnalytics();
+
+  const menuLinks: Omit<MenuListItemProps, "currentPath">[] = [
+    {
+      href: getTeachersUrl(),
+      target: "_blank",
+      onClick: () => track.teacherHubSelected({ navigatedFrom: "menu" }),
+      linkText: "Teacher hub",
+      fontFamily: "heading",
+      fontSize: [32],
+      $mt: [20],
+      arrowSize: [48],
+    },
+    {
+      href: getPupilsUrl(),
+      target: "_blank",
+      onClick: () => track.teacherHubSelected({ navigatedFrom: "menu" }),
+      linkText: "Classroom",
+      fontSize: [32],
+      fontFamily: "heading",
+      $mt: [16],
+      arrowSize: [48],
+    },
+    {
+      href: "/lesson-planning",
+      linkText: "Plan a lesson",
+      fontSize: [24],
+      fontFamily: "heading",
+      $mt: [32],
+      arrowSize: [30],
+    },
+    {
+      href: "/develop-your-curriculum",
+      linkText: "Develop your curriculum",
+      fontSize: [24],
+      fontFamily: "heading",
+      $mt: [12],
+      arrowSize: [30],
+    },
+    {
+      href: "/blog",
+      linkText: "Blog",
+      fontSize: [16],
+      fontFamily: "ui",
+      $mt: [32],
+      arrowSize: [20],
+    },
+    {
+      href: "/about-us/who-we-are",
+      linkText: "About us",
+      fontSize: [16],
+      fontFamily: "ui",
+      $mt: [8],
+      arrowSize: [20],
+    },
+    {
+      href: "/contact-us",
+      linkText: "Contact us",
+      fontSize: [16],
+      fontFamily: "ui",
+      $mt: [8],
+      arrowSize: [20],
+    },
+    {
+      href: getHelpUrl(),
+      linkText: "Help",
+      fontSize: [16],
+      fontFamily: "ui",
+      $mt: [8],
+      arrowSize: [20],
+    },
+  ];
 
   return (
     <FixedHeader $background={theme.header.background}>
@@ -34,11 +105,27 @@ const SiteHeader: FC = () => {
       </Link>
       <Flex $alignItems={"center"} $display={["none", "flex"]} $ml={["auto"]}>
         <P>
-          <Link href={"https://classroom.thenational.academy/"}>Classroom</Link>
+          <Link href={getPupilsUrl()}>
+            <a
+              onClick={() =>
+                track.classroomSelected({ navigatedFrom: "header" })
+              }
+              target="_blank"
+            >
+              Classroom
+            </a>
+          </Link>
         </P>
         <P $ml={24} $mr={32}>
-          <Link href={"https://teachers.thenational.academy/"}>
-            Teacher Hub
+          <Link href={getTeachersUrl()}>
+            <a
+              onClick={() =>
+                track.teacherHubSelected({ navigatedFrom: "header" })
+              }
+              target="_blank"
+            >
+              Teacher Hub
+            </a>
           </Link>
         </P>
       </Flex>
@@ -52,75 +139,7 @@ const SiteHeader: FC = () => {
         }}
       />
       <Menu>
-        {/* TODO:
-         * move menu contents into new component using fixture
-         * add urls for new pages
-         * add moving arrow
-         */}
-        <Flex $alignItems={"center"}>
-          <LocationIcon variant="minimal" name="ArrowRight" size={[48]} />
-          <Home tag={"h2"} $fontSize={[32]} $color={"black"}>
-            <Link href={"/"}>Home</Link>
-          </Home>
-        </Flex>
-        <ul role="list">
-          <li>
-            <P $fontFamily={"heading"} $fontSize={[32]} $mt={[20]}>
-              <Link href={"https://teachers.thenational.academy/"}>
-                Teacher Hub
-              </Link>
-            </P>
-          </li>
-          <li>
-            <P $fontFamily={"heading"} $fontSize={[32]} $mt={[16]}>
-              <Link href={"https://classroom.thenational.academy/"}>
-                Classroom
-              </Link>
-            </P>
-          </li>
-          <li>
-            <P $fontFamily={"heading"} $fontSize={[24]} $mt={[32]}>
-              <Link href={"/develop-your-curriculum"}>
-                Develop Your Curriculum
-              </Link>
-            </P>
-          </li>
-          <li>
-            <P $fontFamily={"heading"} $fontSize={[24]} $mt={[12]}>
-              <Link href={"/support-your-team"}>Support Your Team</Link>
-            </P>
-          </li>
-          <li>
-            <P $fontFamily={"heading"} $fontSize={[24]} $mt={[12]}>
-              <Link href={"/lesson-planning"}>Plan a lesson</Link>
-            </P>
-          </li>
-          <li>
-            <P $fontFamily={"ui"} $fontSize={[16]} $mt={[32]}>
-              <Link href={"/blog"}>Blogs</Link>
-            </P>
-          </li>
-          <li>
-            <P $fontFamily={"ui"} $fontSize={[16]} $mt={[8]}>
-              <Link href={"/webinars"}>Webinars</Link>
-            </P>
-          </li>
-          <li>
-            <P $fontFamily={"ui"} $fontSize={[16]} $mt={[8]}>
-              <Link href={"/about-us/who-we-are"}>About us</Link>
-            </P>
-          </li>
-          <li>
-            <P $fontFamily={"ui"} $fontSize={[16]} $mt={[8]}>
-              <Link href={"/contact-us"}>Contact us</Link>
-            </P>
-          </li>
-          <li>
-            <P $fontFamily={"ui"} $fontSize={[16]} $mt={[8]}>
-              <Link href={"https://support.thenational.academy/"}>Help</Link>
-            </P>
-          </li>
-        </ul>
+        <MenuLinks menuLinks={menuLinks} currentPath={asPath} />
       </Menu>
     </FixedHeader>
   );