+++ src/node-lib/cms/types/documents.ts
@@ -1,6 +1,4 @@
-import { Image } from "../../sanity-graphql/generated/sdk";
-
-import { Document, PortableTextJSON } from "./base";
+import { Document, PortableTextJSON, SanityImage } from "./base";
 
 export type TeamMember = Document & {
   name: string;
@@ -8,32 +6,34 @@ export type TeamMember = Document & {
 
 export type TeamMemberPreview = Pick<TeamMember, "name">;
 
+export type BlogWebinarCategory = {
+  title: string;
+  slug: string;
+};
+
 export type Webinar = Document & {
   title: string;
   slug: string;
   date: Date;
   hosts: TeamMember[];
+  category: BlogWebinarCategory;
   summaryPortableText: PortableTextJSON;
 };
 
 export type WebinarPreview = Pick<
   Webinar,
-  "id" | "title" | "slug" | "summaryPortableText"
+  "id" | "title" | "slug" | "date" | "category" | "summaryPortableText"
 >;
 
 export type BlogPost = Document & {
   title: string;
   slug: string;
   date: Date;
-  mainImage: Image;
+  mainImage: SanityImage;
   author: TeamMember;
+  summary: string;
+  category: BlogWebinarCategory;
   contentPortableText: PortableTextJSON;
-  category: {
-    title: string;
-    slug: {
-      current: string;
-    };
-  };
 };
 
 export type BlogPostPreview = Pick<
@@ -41,6 +41,7 @@ export type BlogPostPreview = Pick<
   | "id"
   | "title"
   | "slug"
+  | "summary"
   | "contentPortableText"
   | "author"
   | "category"