+++ src/node-lib/cms/types/pages.ts
@@ -82,7 +82,10 @@ export type CurriculumPage = Document & {
   gettingStarted: TextBlock;
   elements: {
     title: string;
-    posts: { post: { title: string; slug: { current: string } } }[];
+    posts: {
+      title: string;
+      post: { title: string; slug: { current: string } };
+    }[];
   };
   ourApproach: TextBlock;
 };