+++ src/pages/blog/index.tsx
@@ -4,15 +4,17 @@ import { DEFAULT_SEO_PROPS } from "../../browser-lib/seo/Seo";
 import BlogList from "../../components/BlogList";
 import { BlogListItemProps } from "../../components/BlogList/BlogListItem";
 import Layout from "../../components/Layout";
-import CMSClient from "../../node-lib/cms";
+import CMSClient, { BlogPostPreview } from "../../node-lib/cms";
 import MaxWidth from "../../components/MaxWidth/MaxWidth";
 import Grid, { GridArea } from "../../components/Grid";
 import SummaryCard from "../../components/Card/SummaryCard";
 
-import { SerializedBlog } from "./[blogSlug]";
+export type SerializedBlogPostPreview = Omit<BlogPostPreview, "date"> & {
+  date: string;
+};
 
 export type BlogListingPageProps = {
-  blogs: SerializedBlog[];
+  blogs: SerializedBlogPostPreview[];
   isPreviewMode: boolean;
 };
 
@@ -21,7 +23,7 @@ const BlogListingPage: NextPage<BlogListingPageProps> = (props) => {
   console.log(props);
 
   const cardImage = {
-    imageSrc: "/images/illustrations/teacher-carrying-stuff.png",
+    src: "/images/illustrations/teacher-carrying-stuff.png",
     alt: "",
   };
 
@@ -40,7 +42,7 @@ const BlogListingPage: NextPage<BlogListingPageProps> = (props) => {
             "Read blogs from our in-house experts to find ideas to take away and try, from curriculum planning to lesson delivery. Plus, keep up to date with the latest news and insights from Oak."
           }
           background="teachersPastelYellow"
-          cardImageProps={cardImage}
+          imageProps={cardImage}
         />
         <Grid>
           <GridArea $colSpan={[12, 12, 7]} $mt={[48, 72]}>
@@ -56,16 +58,24 @@ const BlogListingPage: NextPage<BlogListingPageProps> = (props) => {
   );
 };
 
-const blogToBlogListItem = (blog: SerializedBlog): BlogListItemProps => ({
+export const blogToBlogListItem = (
+  blog: SerializedBlogPostPreview
+): BlogListItemProps => ({
   contentType: "blog-post",
   title: blog.title,
   href: `/blog/${blog.slug}`,
-  snippet:
-    "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. ",
+  snippet: blog.summary,
   titleTag: "h3",
-  category: blog.category.title,
+  category: blog.category,
   date: blog.date,
-  mainImage: blog?.mainImage?.asset?.url || "",
+  mainImage: blog?.mainImage,
+});
+
+export const serializeDate = <T extends { date: Date }>(
+  item: T
+): T & { date: string } => ({
+  ...item,
+  date: item.date.toISOString(),
 });
 
 export const getStaticProps: GetStaticProps<BlogListingPageProps> = async (
@@ -77,12 +87,7 @@ export const getStaticProps: GetStaticProps<BlogListingPageProps> = async (
     previewMode: isPreviewMode,
   });
 
-  const blogs = blogResults.map((blog) => {
-    return {
-      ...blog,
-      date: blog.date.toISOString(),
-    };
-  });
+  const blogs = blogResults.map(serializeDate);
 
   return {
     props: {